isis: extract tilfa.rs from inst.rs (PR 1 of 4)
First slice of the inst.rs split per the Proposal B plan
(tilfa → graph → lsp → rib). Pure code motion — no behavior change.

Moved to new `isis/tilfa.rs`:
  - struct RepairPathMpls / RepairPathSrv6
  - pub(super) fn first_router_hop_id
  - pub(super) fn build_repair_path_mpls
  - pub(super) fn build_repair_path_srv6
  - pub(super) fn tilfa_repair_path
  - private SR-label resolution chain (enum SrBlockKind,
    resolve_sid_to_label, node_sid_label_for_vertex,
    adj_sid_label_for_link, repair_segments_to_mpls_labels)

Visibility flips:
  - first_router_hop_id / build_repair_path_* / tilfa_repair_path:
    fn -> pub(super) so inst.rs callers (build_rib_from_spf*,
    perform_spf_calculation) can reach them.
  - RepairPathMpls / RepairPathSrv6: stay pub (referenced by
    SpfNexthop.backup / SpfNexthopV6.backup fields in inst.rs and
    by the detail renderer in show.rs).
  - SR-label-resolution chain: stay private to tilfa.rs.

Imports updated:
  - inst.rs: `use super::tilfa::{...}` for what it now calls
    cross-module; EncapType moved behind `#[cfg(test)]` since only
    the make_rib_entry_v6 test fixture references it directly.
  - show.rs: RepairPathMpls/Srv6 imported from `super::tilfa::`
    instead of `super::inst::`.
  - mod.rs: registers `pub mod tilfa;` (no re-export — only sibling
    modules in isis/ need these types).

inst.rs: 3610 -> 3295 lines (-315). tilfa.rs: 340 lines.
342 tests pass; clippy clean.
